Clear and unify modules

+++ module03/ex00/src/main.cpp
@@ -1,16 +1,22 @@
+// @author    Clément Vidon
+// @created   230324 12:56:23  by  clem@spectre
+// @modified  230329 15:58:10  by  clem@spectre
+// @filename  main.cpp
+
 #include <cstdlib>
 #include <iostream>
+
 #include "ClapTrap.hpp"
 
 int main( void ) {
-  int      random;
-  ClapTrap ct0;
-  ClapTrap ct1( "Charly" );
+  unsigned int random;
+  ClapTrap     ct0;
+  ClapTrap     ct1( "Charly" );
 
-  srand( time( NULL ) );
+  srand( static_cast<unsigned int>( time( NULL ) ) );
   std::cout << ">> Start <<" << std::endl;
   while( ct0.isAble() && ct1.isAble() ) {
-    random = rand() % 100 + 1;
+    random = static_cast<unsigned int>( rand() % 100 + 1 );
 
     if( random <= 50 ) {
       ct0.attack( ct1.getName() );
@@ -20,7 +26,7 @@ int main( void ) {
       ct0.takeDamage( ct1.getAttackDamage() );
     }
 
-    random = rand() % 100 + 1;
+    random = static_cast<unsigned int>( rand() % 100 + 1 );
     if( random <= 50 ) {
       ct0.beRepaired( random % 10 );
     } else {
